fonction learning create_scene mieux implémenté

+++ main/background.js
@@ -1,19 +1,8 @@
-import {
-  app,
-  protocol,
-  BrowserWindow,
-  ipcMain,
-  Menu,
-  dialog,
-  session
-} from "electron"
+import { app, protocol, BrowserWindow, ipcMain, Menu, dialog, session } from "electron"
 import axios from "axios"
 import serve from "electron-serve"
 import { createWindow } from "./helpers"
-import {
-  installExtension,
-  REACT_DEVELOPER_TOOLS
-} from "electron-extension-installer"
+import { installExtension, REACT_DEVELOPER_TOOLS } from "electron-extension-installer"
 const fs = require("fs")
 var path = require("path")
 const os = require("node:os")
@@ -22,7 +11,7 @@ var serverProcess = null
 var flaskPort = 5000
 var hasBeenSet = false
 
-const RUN_SERVER_WITH_APP = true
+const RUN_SERVER_WITH_APP = false
 
 const isProd = process.env.NODE_ENV === "production"
 
@@ -31,11 +20,6 @@ if (isProd) {
 } else {
   app.setPath("userData", `${app.getPath("userData")} (development)`)
 }
-
-const reactDevToolsPath = path.join(
-  os.homedir(),
-  "\\AppData\\Local\\Google\\Chrome\\User Data\\Default\\Extensions\\fmkadmapgofadopljbjfkapdkoienihi\\4.28.0_0"
-)
 
 ;(async () => {
   await app.whenReady()
@@ -79,14 +63,7 @@ const reactDevToolsPath = path.join(
     },
     {
       label: "Edit",
-      submenu: [
-        { role: "undo" },
-        { role: "redo" },
-        { type: "separator" },
-        { role: "cut" },
-        { role: "copy" },
-        { role: "paste" }
-      ]
+      submenu: [{ role: "undo" }, { role: "redo" }, { type: "separator" }, { role: "cut" }, { role: "copy" }, { role: "paste" }]
     },
     {
       label: "Hello From Electron!",
@@ -111,19 +88,12 @@ const reactDevToolsPath = path.join(
   ]
 
   // link: https://medium.com/red-buffer/integrating-python-flask-backend-with-electron-nodejs-frontend-8ac621d13f72
-  console.log(
-    RUN_SERVER_WITH_APP
-      ? "Server will start automatically here (in background of the application)"
-      : "Server must be started manually"
-  )
+  console.log(RUN_SERVER_WITH_APP ? "Server will start automatically here (in background of the application)" : "Server must be started manually")
   if (RUN_SERVER_WITH_APP) {
     if (!isProd) {
       //**** DEVELOPMENT ****//
       // IMPORTANT: Select python interpreter (related to your virtual environment)
-      var path2conda = fs.readFileSync(
-        "./path2condaenv_toDeleteInProd.txt",
-        "utf8"
-      )
+      var path2conda = fs.readFileSync("./path2condaenv_toDeleteInProd.txt", "utf8")
 
       const net = require("net")
 
@@ -162,10 +132,7 @@ const reactDevToolsPath = path.join(
       findAvailablePort(5000, 8000)
         .then((port) => {
           console.log(`Available port: ${port}`)
-          serverProcess = require("child_process").spawn(path2conda, [
-            "./flask_server/server.py",
-            "--port=" + port
-          ])
+          serverProcess = require("child_process").spawn(path2conda, ["./flask_server/server.py", "--port=" + port])
           flaskPort = port
           serverProcess.stdout.on("data", function (data) {
             console.log("data: ", data.toString("utf8"))
@@ -254,19 +221,6 @@ const reactDevToolsPath = path.join(
     mainWindow.webContents.openDevTools()
   }
 })()
-// .then(async () => {
-//   await session.defaultSession.loadExtension(reactDevToolsPath)
-// })
-
-// .then(async () => {
-//   await session.defaultSession.loadExtension(
-//     path.join(__dirname, "react-devtools"),
-//     // allowFileAccess is required to load the devtools extension on file:// URLs.
-//     { allowFileAccess: true }
-//   )
-//   // Note that in order to use the React DevTools extension, you'll need to
-//   // download and unzip a copy of the extension.
-// })
 
 /**
  * @description Set the working directory
@@ -292,10 +246,7 @@ function setWorkingDirectory(event, mainWindow) {
         if (file === app.getPath("sessionData")) {
           // If the working directory is already set to the selected folder
           console.log("Working directory is already set to " + file)
-          event.reply(
-            "messageFromElectron",
-            "Working directory is already set to " + file
-          )
+          event.reply("messageFromElectron", "Working directory is already set to " + file)
           event.reply("workingDirectorySet", {
             workingDirectory: dirTree(file),
             hasBeenSet: hasBeenSet
@@ -364,7 +315,7 @@ ipcMain.handle("request", async (_, axios_request) => {
 app.on("window-all-closed", () => {
   app.quit()
   console.log("app quit")
-  if (!isProd) {
+  if (!isProd && RUN_SERVER_WITH_APP) {
     serverProcess.kill()
     console.log("serverProcess killed")
   }
